requirements.txt
flake8
nltk
pandas
python-decouple
tweepy
